Propagate all agent labels, adjust test, refactor
