+++ internal/webhooks/agent/test/resources.go
@@ -79,10 +79,19 @@ func (r *AgentWebhookTestResources) NewDeployment() *appsv1.Deployment {
 			Name:      r.Name,
 			Namespace: r.Namespace,
 			Labels: map[string]string{
-				"app":                   r.Name,
-				"other":                 "label",
-				"cryostat.io/namespace": r.Namespace,
-				"cryostat.io/name":      "cryostat",
+				"app":                                 r.Name,
+				"other":                               "label",
+				"cryostat.io/namespace":               r.Namespace,
+				"cryostat.io/name":                    "cryostat",
+				"cryostat.io/log-level":               "info",
+				"cryostat.io/callback-port":           "123",
+				"cryostat.io/container":               "other-container",
+				"cryostat.io/read-only":               "false",
+				"cryostat.io/java-options-var":        "somevar",
+				"cryostat.io/harvester-template":      "all",
+				"cryostat.io/harvester-exit-max-age":  "456s",
+				"cryostat.io/harvester-exit-max-size": "789",
+				"cryostat.io/smart-triggers":          "someConfigMap",
 			},
 			Annotations: map[string]string{
 				"app.openshift.io/connects-to": "something-else",
@@ -124,10 +133,19 @@ func (r *AgentWebhookTestResources) NewMutatedDeployment() *appsv1.Deployment {
 			Name:      r.Name,
 			Namespace: r.Namespace,
 			Labels: map[string]string{
-				"app":                   "something-else",
-				"other":                 "label",
-				"cryostat.io/namespace": r.Namespace,
-				"cryostat.io/name":      "cryostat",
+				"app":                                 "something-else",
+				"other":                               "label",
+				"cryostat.io/namespace":               r.Namespace,
+				"cryostat.io/name":                    "cryostat",
+				"cryostat.io/log-level":               "info",
+				"cryostat.io/callback-port":           "123",
+				"cryostat.io/container":               "other-container",
+				"cryostat.io/read-only":               "false",
+				"cryostat.io/java-options-var":        "somevar",
+				"cryostat.io/harvester-template":      "all",
+				"cryostat.io/harvester-exit-max-age":  "456s",
+				"cryostat.io/harvester-exit-max-size": "789",
+				"cryostat.io/smart-triggers":          "someConfigMap",
 			},
 			Annotations: map[string]string{
 				"app.openshift.io/connects-to": "something-else",
@@ -140,9 +158,18 @@ func (r *AgentWebhookTestResources) NewMutatedDeployment() *appsv1.Deployment {
 					Name:      r.Name,
 					Namespace: r.Namespace,
 					Labels: map[string]string{
-						"app":                   "someApplication",
-						"cryostat.io/namespace": r.Namespace,
-						"cryostat.io/name":      "cryostat",
+						"app":                                 "someApplication",
+						"cryostat.io/namespace":               r.Namespace,
+						"cryostat.io/name":                    "cryostat",
+						"cryostat.io/log-level":               "info",
+						"cryostat.io/callback-port":           "123",
+						"cryostat.io/container":               "other-container",
+						"cryostat.io/read-only":               "false",
+						"cryostat.io/java-options-var":        "somevar",
+						"cryostat.io/harvester-template":      "all",
+						"cryostat.io/harvester-exit-max-age":  "456s",
+						"cryostat.io/harvester-exit-max-size": "789",
+						"cryostat.io/smart-triggers":          "someConfigMap",
 					},
 				},
 				Spec: corev1.PodSpec{
